% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utilities.R
\name{add_group_key}
\alias{add_group_key}
\title{Help for multiple variable grouping}
\usage{
add_group_key(data, group_vars, key_name = "..group_id")
}
\arguments{
\item{data}{dataset.}

\item{group_vars}{variables used for grouping (repreated measurements)}

\item{key_name}{column name for the created variable}
}
\description{
Help for multiple variable grouping
}
