% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_functions.R
\name{apply_k_extended}
\alias{apply_k_extended}
\title{Extended K-anonymity}
\usage{
apply_k_extended(df, qids, k, bucket_cols = list(), direct_ids = character())
}
\description{
Extended K-anonymity
}
