% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anonymization_functions.R
\name{apply_masking}
\alias{apply_masking}
\title{Masking}
\usage{
apply_masking(df, cols)
}
\arguments{
\item{df}{is data frame}

\item{cols}{the selected column name}
}
\description{
This fuction uses asterics to partially or wholly hide a section of variable
}
