% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGBM.R
\name{autoGBM}
\alias{autoGBM}
\title{autoGBM}
\usage{
autoGBM(x, y, train_hex, valid_hex, test_hex, model_path = "./")
}
\arguments{
\item{x}{independent variables}

\item{y}{dependent variable}

\item{data_hex}{H2ODataFrame}
}
\description{
auto train GBM using pre-defined strategies
}
\examples{
library(h2o)
library(rAutoML)
model_path <- c("C:/tmp")
h2o.init()
data(churn, package = "rAutoML")
churn_hex <- as.h2o(churn)
split_hex <- h2o.splitFrame(data = churn_hex, ratios = c(0.5,0.3), seed = 1234)
train_hex <- split_hex[[1]]
valid_hex <- split_hex[[2]]
test_hex  <- split_hex[[3]]
y <- "Churn."
x <- setdiff(names(churn_hex),  c(y))
autoGBM(x, y, train_hex, valid_hex, test_hex, model_path)
}
