% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGBM.R
\name{bayes_gbm}
\alias{bayes_gbm}
\title{bayes_gbm}
\usage{
bayes_gbm(max_depth, learn_rate, sample_rate, col_sample_rate)
}
\arguments{
\item{max_depth}{max_depth}

\item{learn_rate}{learn_rate}

\item{sample_rate}{sample_rate}

\item{col_sample_rate}{col_sample_rate}
}
\description{
gbm function for bayesian grid search
}
