% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{boot_estimates}
\alias{boot_estimates}
\title{Bootstrap estimate for the predictive measures}
\usage{
boot_estimates(
  model,
  df,
  outcome_var,
  problem_type,
  nreps = 100,
  type = "prob",
  model_name = NULL,
  report = c("Accuracy", "AUCROC", "AUCRecall", "Sens", "Spec", "Precision", "Recall",
    "F", "RMSE", "Rsquared", "MAE"),
  summary_fun = quantile_summary
)
}
\description{
Bootstrap estimate for the predictive measures
}
