% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{boot_estimates_multiple.caretEnsemble}
\alias{boot_estimates_multiple.caretEnsemble}
\title{Compute predicitve metrics for multiple caretEnsemble models}
\usage{
\method{boot_estimates_multiple}{caretEnsemble}(
  models,
  df,
  outcome_var,
  problem_type,
  nreps = 100,
  model_name = NULL,
  type = "prob",
  report = c("Accuracy", "AUCROC", "AUCRecall", "Sens", "Spec", "Precision", "Recall",
    "F", "RMSE", "Rsquared", "MAE"),
  summary_fun = quantile_summary
)
}
\description{
Compute predicitve metrics for multiple caretEnsemble models
}
