% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{boot_estimates_multiple}
\alias{boot_estimates_multiple}
\title{Compute predicitve metrics for multiple caret models}
\usage{
boot_estimates_multiple(
  models,
  df,
  outcome_var,
  problem_type,
  nreps,
  model_name,
  type,
  report,
  summary_fun,
  save_model,
  model_id,
  model_folder,
  recipe_folder,
  preprocesses
)
}
\description{
Compute predicitve metrics for multiple caret models
}
