% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{boot_measures}
\alias{boot_measures}
\title{Estimate predictive scores for a single model}
\usage{
boot_measures(model, df, outcome_var, problem_type, type = "prob")
}
\arguments{
\item{df}{test data.frame}
}
\description{
Estimate predictive scores for a single model
}
