% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{boxplot_custom}
\alias{boxplot_custom}
\title{Generate a Boxplot}
\usage{
boxplot_custom(
  df,
  xvar = NULL,
  yvar,
  colorVar = NULL,
  plot_title = NULL,
  vertical = TRUE,
  xlab = NULL,
  ylab = NULL,
  title_pos = 0.5,
  title_size = 25,
  default_col = "#0077B6",
  axis_title_size = 20,
  axis_text_size = 16,
  axistext_angle = 0,
  legend_title = "",
  colorbrewer = "Dark2"
)
}
\description{
Generate a Boxplot
}
