% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_api_connection}
\alias{check_api_connection}
\title{Check if API connection is created}
\usage{
check_api_connection(url, timeout = 5)
}
\arguments{
\item{url}{}
}
\description{
Check if API connection is created
}
