% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_columns}
\alias{check_columns}
\title{Check if columns in two datasets match}
\usage{
check_columns(df1, df2)
}
\arguments{
\item{df1}{first data frame.}

\item{df2}{second data frame.}
}
\description{
Check if columns in two datasets match
}
