% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_value_exists}
\alias{check_value_exists}
\title{Check a value within a data frame}
\usage{
check_value_exists(df, var, what)
}
\arguments{
\item{df}{data frame}

\item{var}{variable to search from}

\item{what}{character string to search}
}
\description{
Check a value within a dataframe
}
