% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{collect_logs}
\alias{collect_logs}
\title{Collect log files}
\usage{
collect_logs(path, pattern = "*.log$")
}
\arguments{
\item{path}{path to the files.}

\item{pattern}{search pattern.}
}
\value{
dataframe object.
}
\description{
Create a data frame of log files based on a particular pattern.
}
\details{
Collect all log files and create a dataframe.
}
