% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_data}
\alias{combine_data}
\title{Add rows or columns}
\usage{
combine_data(df1, df2, type, id = NULL)
}
\arguments{
\item{df1}{base dataset}

\item{df2}{new dataset}

\item{type}{either "row" or "column"}

\item{id}{indicator variable}
}
\description{
Add rows or columns
}
\details{
Combine two datasets by row, i.e., append  or by columns
}
