% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{combine_rows}
\alias{combine_rows}
\title{Combine rows}
\usage{
combine_rows(df1, df2, id = NULL)
}
\arguments{
\item{df1}{base dataset}

\item{df2}{new dataset}

\item{id}{indicator variable}
}
\description{
Combine rows
}
\details{
Combine two datasets by row, i.e., append
}
