% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{combine_shap_plots}
\alias{combine_shap_plots}
\title{Collect SHAP viz plots}
\usage{
combine_shap_plots(vv, type)
}
\arguments{
\item{vv}{compute_shap object with named models}

\item{type}{various sv objets, either "sv_bw",  "sv_wf" or "sv_f"}
}
\description{
Collect SHAP viz plots
}
