% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_shap_values.R
\name{compute_shap.caretList}
\alias{compute_shap.caretList}
\title{Compute SHAP values for several caretList models}
\usage{
\method{compute_shap}{caretList}(
  models,
  model_names,
  newdata,
  response,
  task,
  nsim = 50,
  max_n = 1000,
  top_n_rank = 5,
  total_n_rank = 20
)
}
\arguments{
\item{models}{caretList models}

\item{model_names}{names of the models as in the \code{names(models)}.}

\item{newdata}{training data}
}
\description{
Compute SHAP values for several caretList models
}
