% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{compute_shap}
\alias{compute_shap}
\title{Compute SHAP values for several caretEnsemble models}
\usage{
compute_shap(
  models,
  model_names = NULL,
  newdata,
  response,
  task,
  nsim = 50,
  max_n = 1000,
  top_n_rank = 5,
  total_n_rank = 20
)
}
\description{
Compute SHAP values for several caretEnsemble models
}
