% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_shap_values.R
\name{compute_shap_single}
\alias{compute_shap_single}
\title{Compute SHAP values for a caretList or caretEnsemble model}
\usage{
compute_shap_single(
  model,
  newdata,
  response,
  task,
  positive_class = 2,
  model_name = "model",
  nsim = 50,
  max_n = 1000
)
}
\description{
Compute SHAP values for a caretList or caretEnsemble model
}
\details{
Computes SHAP values for a single caret model. In case of cartList or caretEnsemble, call the specific model from the list.
}
