% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_database.R
\name{connect_to_database}
\alias{connect_to_database}
\title{Connect to a database}
\usage{
connect_to_database(dbms, server, user, password, port, pathToDriver)
}
\arguments{
\item{dbms}{The type of DBMS running on the server}

\item{server}{The name of the server}

\item{user}{The user name used to access the server}

\item{password}{The password for that user}

\item{port}{The port number on the server}

\item{pathToDriver}{Path to the JDBC driver (optional)}
}
\value{
A list containing the connection details and connection object
}
\description{
This function establishes a connection to a database using the specified parameters.
}
\examples{
\dontrun{
connect_to_database(
  dbms = "postgresql",
  server = "localhost/db",
  user = "user",
  password = "password",
  port = 5432
)
}
}
