% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cdm_reference.R
\name{create_cdm_reference}
\alias{create_cdm_reference}
\title{Create a CDM reference}
\usage{
create_cdm_reference(
  con,
  cdmName,
  cdmSchema,
  writeSchema,
  achillesSchema = NULL
)
}
\arguments{
\item{con}{A database connection object}

\item{cdmName}{The name of the CDM}

\item{cdmSchema}{The schema containing the CDM tables}

\item{writeSchema}{The schema where results can be written}

\item{achillesSchema}{Optional schema for Achilles results}
}
\value{
A CDM reference object
}
\description{
This function creates a CDM reference using CDMConnector.
}
\examples{
\dontrun{
cdm <- create_cdm_reference(con, "my_cdm", "cdm_schema", "results_schema")
}
}
