% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_df_metadata}
\alias{create_df_metadata}
\title{Create dataset metadata}
\usage{
create_df_metadata(
  data,
  filename,
  study_name,
  study_country,
  additional_info,
  upload_time = Sys.time(),
  last_modified = upload_time,
  user = "Admin",
  groups = "admin"
)
}
\arguments{
\item{data}{uploaded dataset.}

\item{filename}{a character string. Can be file name of the data or a file path, with the correct file extension.}

\item{study_name}{a short description about the study or analysis.}

\item{study_country}{Country where the study was conducted or where the data was collected from.}

\item{upload_time}{timestamp indicating when the data was uploaded.}

\item{user}{user accessing the platform}

\item{group}{collaborating groups which have access to the data.}
}
\value{
a dataframe.
}
\description{
Create a file with the information about the data.
}
\details{
The function collects user provided information, together with the system generate to create a metadata file.
}
