% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_dir}
\alias{create_dir}
\title{Create directory for various artifacts}
\usage{
create_dir(name = "new_folder")
}
\arguments{
\item{name}{a directory name or relative path.}
}
\description{
Create directory for various artifacts
}
\details{
Checks if the directory exists otherwise creates it.
}
