% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_models.R
\name{create_ensemble.caretList}
\alias{create_ensemble.caretList}
\title{Create ensemble model}
\usage{
\method{create_ensemble}{caretList}(
  all.models,
  excluded_class_id = 0L,
  tuneLength = 100,
  ctrl = NULL,
  metric = NULL
)
}
\description{
Create ensemble model
}
