% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{create_ensemble}
\alias{create_ensemble}
\title{Create ensemble model}
\usage{
create_ensemble(
  all.models,
  excluded_class_id = 0L,
  tuneLength = 100,
  ctrl = NULL,
  metric = NULL
)
}
\description{
Create ensemble model
}
