% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utilities.R
\name{create_grouped_index}
\alias{create_grouped_index}
\title{Create index for grouped k-fold cross-validation}
\usage{
create_grouped_index(x, k = 5)
}
\arguments{
\item{x}{vector of groups in the data}

\item{k}{k-fold}
}
\description{
Create index for grouped k-fold cross-validation
}
