% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_model_logs}
\alias{create_model_logs}
\title{Create trained model logs}
\usage{
create_model_logs(
  df_name,
  session_name,
  outcome,
  framework,
  train_result,
  timestamp = Sys.time(),
  path = ".log_files"
)
}
\description{
Create trained model logs
}
