% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{create_pred_prototype}
\alias{create_pred_prototype}
\title{Generate prediction prototype data}
\usage{
create_pred_prototype(df, outcome_var = NULL)
}
\arguments{
\item{df}{data.frame used to train the model}

\item{outcome_var}{outcome variable}
}
\description{
Generate prediction prototype data
}
