% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{create_prompts}
\alias{create_prompts}
\title{Create prompts}
\usage{
create_prompts(
  desc = "Help me describe the R output: ",
  log,
  add_info = "Write in paragraph and provide details"
)
}
\arguments{
\item{desc}{prompt for the GPT}

\item{log}{list object}

\item{add_info}{additional infomation for the prompt}
}
\description{
Create all the prompts for GPT
}
