% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_barplot.R
\name{custom_barplot}
\alias{custom_barplot}
\title{Bar Plot Function}
\usage{
custom_barplot(
  df,
  xvar,
  yvar = NULL,
  vertical = TRUE,
  stackedtype = TRUE,
  colorVar = NULL,
  bar_width = 0.9,
  default_col = "blue",
  plot_title = NULL,
  xlab = NULL,
  ylab = "Percentage",
  title_pos = 0.5,
  title_size = 25,
  axis_title_size = 20,
  axis_text_size = 16,
  data_label_size = 10,
  axistext_angle = 0,
  legend_title = "",
  colorbrewer = "Dark2",
  custom_theme = theme_minimal()
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{xvar}{The categorical variable for the x-axis.}

\item{yvar}{The numerical variable for the y-axis (optional).}

\item{vertical}{Logical; whether the bars should be vertical (default: TRUE).}

\item{stackedtype}{Logical; whether the bars should be stacked (default: TRUE).}

\item{colorVar}{A categorical variable for coloring bars (optional).}

\item{bar_width}{Width of the bars (default: 0.9).}

\item{default_col}{Default fill color if colorVar is not provided (default: "blue").}

\item{plot_title}{Title of the plot (optional).}

\item{xlab}{Label for the x-axis (optional).}

\item{ylab}{Label for the y-axis (default: "Percentage").}

\item{title_pos}{Positioning of the title (default: 0.5).}

\item{title_size}{Size of the title text (default: 25).}

\item{axis_title_size}{Size of the axis titles (default: 20).}

\item{axis_text_size}{Size of the axis text (default: 16).}

\item{data_label_size}{Size of the data labels (default: 10).}

\item{axistext_angle}{Angle of x-axis text labels (default: 0).}

\item{legend_title}{Title of the legend (default: "").}

\item{colorbrewer}{Color palette from RColorBrewer (default: "Dark2").}
}
\value{
A ggplot2 bar plot object.
}
\description{
Generates a bar plot with various customization options.
}
