% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_boxplot.R
\name{custom_boxplot}
\alias{custom_boxplot}
\title{Customizable Boxplot Function}
\usage{
custom_boxplot(
  df,
  xvar = NULL,
  yvar,
  colorVar = NULL,
  plot_title = NULL,
  vertical = TRUE,
  xlab = NULL,
  ylab = NULL,
  title_pos = 0.5,
  title_size = 25,
  default_col = "#0077B6",
  axis_title_size = 20,
  axis_text_size = 16,
  axistext_angle = 0,
  legend_title = "",
  colorbrewer = "Dark2",
  custom_theme = theme_minimal()
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{xvar}{Optional. A categorical variable for the x-axis.}

\item{yvar}{Required. A numerical variable for the y-axis.}

\item{colorVar}{Optional. A categorical variable for color grouping.}

\item{plot_title}{Optional. Title for the plot.}

\item{vertical}{Logical. If TRUE (default), plots a vertical boxplot; otherwise, a horizontal one.}

\item{xlab}{Optional. Label for the x-axis.}

\item{ylab}{Optional. Label for the y-axis.}

\item{title_pos}{Numeric. Position of the title (default is 0.5).}

\item{title_size}{Integer. Font size of the title (default is 25).}

\item{default_col}{Character. Default fill color (default is "#0077B6").}

\item{axis_title_size}{Integer. Font size for axis titles (default is 20).}

\item{axis_text_size}{Integer. Font size for axis text (default is 16).}

\item{axistext_angle}{Numeric. Angle of x-axis text labels (default is 0).}

\item{legend_title}{Character. Title of the legend (default is "").}

\item{colorbrewer}{Character. Color palette for fill aesthetics (default is "Dark2").}
}
\value{
A ggplot2 object representing the boxplot.
}
\description{
This function creates a customizable boxplot using ggplot2.
}
\examples{
df <- data.frame(Category = rep(c("A", "B"), each = 50), Value = rnorm(100))
boxplot_custom(df, xvar = "Category", yvar = "Value")
}
