% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_corrplot.R
\name{custom_corrplot}
\alias{custom_corrplot}
\title{Custom Correlation Plot using DataExplorer and ColorBrewer}
\usage{
custom_corrplot(df = NULL, features = NULL, colorbrewer = "Dark2")
}
\arguments{
\item{df}{A data frame containing the data.}

\item{features}{A character vector specifying the names of features to include.}

\item{colorbrewer}{A string specifying the ColorBrewer palette to use (default is "Dark2").}
}
\value{
A ggplot2 object representing the correlation plot. If input is invalid, an empty plot is returned.
}
\description{
Generates a correlation plot for selected features of a dataframe,
with a specified ColorBrewer palette applied to a continuous fill scale.
}
\examples{
custom_corrplot(df = mtcars, features = c("mpg", "hp", "wt"), colorbrewer = "YlGnBu")
}
