% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_linegraph.R
\name{custom_linegraph}
\alias{custom_linegraph}
\title{A custom APHRC Line Graph Plotter}
\usage{
custom_linegraph(
  df,
  xvar,
  yvar = NULL,
  colorVar = NULL,
  line_size = 0.9,
  line_type = "solid",
  line_join = "round",
  addlinetype = FALSE,
  plot_title = NULL,
  xlab = NULL,
  ylab = NULL,
  title_pos = 0.5,
  title_size = 25,
  default_col = "#0077B6",
  axis_title_size = 20,
  axis_text_size = 16,
  data_label_size = 10,
  axistext_angle = 0,
  legend_title = "",
  addpoints = FALSE,
  summary_type = "Total",
  colorbrewer = "Dark2",
  custom_theme = theme_minimal()
)
}
\arguments{
\item{df}{A data frame containing the data to be plotted.}

\item{xvar}{The variable to be plotted on the x-axis.}

\item{yvar}{The variable to be plotted on the y-axis (optional).}

\item{colorVar}{A variable to color the lines by (optional).}

\item{line_size}{The size of the line.}

\item{line_type}{The type of the line (e.g., "solid", "dashed").}

\item{line_join}{The type of line join (e.g., "round", "mitre").}

\item{addlinetype}{Whether to add different line types based on the colorVar (default FALSE).}

\item{plot_title}{Title of the plot.}

\item{title_pos}{Position of the title (default 0.5).}

\item{title_size}{Size of the plot title.}

\item{default_col}{Default color for the plot lines.}

\item{axis_title_size}{Size of the axis titles.}

\item{axis_text_size}{Size of the axis text.}

\item{data_label_size}{Size of the data labels.}

\item{axistext_angle}{Angle of the axis text.}

\item{legend_title}{Title of the legend.}

\item{addpoints}{Whether to add points to the line graph (default FALSE).}

\item{summary_type}{Type of summary ("Total" or "Average").}

\item{colorbrewer}{Color palette for the graph (default "Dark2").}

\item{ggplot2::xlab}{Label for the x-axis.}

\item{ggplot2::ylab}{Label for the y-axis.}
}
\value{
A ggplot2::ggplot object of the line graph.
}
\description{
This function plots a flexible line graph that can be customized based on various parameters.
}
