% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_piechart.R
\name{custom_piechart}
\alias{custom_piechart}
\title{Generate a Pie or Doughnut Chart}
\usage{
custom_piechart(
  df,
  xvar,
  plot_title = "Pie Chart",
  title_pos = 0.5,
  title_size = 28,
  facet_var = NULL,
  facet_title_size = 24,
  legend_title = NULL,
  transform_to_doughnut = FALSE,
  data_label_size = 4,
  colorbrewer = "Paired",
  custom_theme = theme_void()
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{xvar}{A string specifying the categorical variable for the pie chart.}

\item{plot_title}{The title of the plot. Default is "Pie Chart".}

\item{title_pos}{Position of the title (default = 0.5).}

\item{title_size}{Size of the title text (default = 28).}

\item{facet_var}{An optional string specifying a facet variable.}

\item{facet_title_size}{Size of facet labels (default = 24).}

\item{legend_title}{Title for the legend (default = NULL).}

\item{transform_to_doughnut}{Logical. If TRUE, creates a doughnut chart (default = FALSE).}

\item{data_label_size}{Size of data labels inside the pie chart (default = 4).}

\item{colorbrewer}{Color palette from RColorBrewer (default = "Paired").}
}
\value{
A ggplot2 pie or doughnut chart.
}
\description{
Generate a Pie or Doughnut Chart
}
