% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_violin.R
\name{custom_violin}
\alias{custom_violin}
\title{Create a Customized Violin Plot}
\usage{
custom_violin(
  df,
  xvar,
  yvar,
  colorVar = NULL,
  plot_title = NULL,
  vertical = TRUE,
  xlab = NULL,
  ylab = NULL,
  title_pos = 0.5,
  title_size = 25,
  default_col = "#0077B6",
  axis_title_size = 20,
  axis_text_size = 16,
  axistext_angle = 0,
  legend_title = "",
  colorbrewer = "Dark2"
)
}
\arguments{
\item{df}{A data frame containing the data.}

\item{xvar}{A string specifying the x-axis variable.}

\item{yvar}{A string specifying the y-axis variable.}

\item{colorVar}{(Optional) A string specifying the color grouping variable.}

\item{plot_title}{(Optional) A string specifying the plot title.}

\item{vertical}{A logical indicating whether the plot should be vertical (default = TRUE).}

\item{xlab}{(Optional) A string specifying the x-axis label.}

\item{ylab}{(Optional) A string specifying the y-axis label.}

\item{title_pos}{A numeric value for the horizontal position of the title (default = 0.5).}

\item{title_size}{An integer specifying the title font size (default = 25).}

\item{default_col}{A string specifying the default fill color (default = "#0077B6").}

\item{axis_title_size}{An integer specifying the axis title font size (default = 20).}

\item{axis_text_size}{An integer specifying the axis text font size (default = 16).}

\item{axistext_angle}{A numeric value specifying the angle of the x-axis text (default = 0).}

\item{legend_title}{A string specifying the legend title (default = "").}

\item{colorbrewer}{A string specifying the color palette from RColorBrewer (default = "Dark2").}
}
\value{
A ggplot object representing the violin plot.
}
\description{
This function generates a violin plot using ggplot2, with options for color, orientation, and titles.
}
