% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{drop_missing_values}
\alias{drop_missing_values}
\title{Drop NAs}
\usage{
drop_missing_values(df, variable = NULL)
}
\arguments{
\item{df}{data frame}

\item{variable.}{Default is NULL. A vector of variable name(s) in the df. If specified, drop NAs for a specific variable otherwise for all the variables in the df}
}
\description{
Drop NAs
}
\details{
Drops NAs for a specific variable or entire dataset
}
