% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{drop_variables}
\alias{drop_variables}
\title{Drop variables}
\usage{
drop_variables(df, vars)
}
\arguments{
\item{df}{data frame}

\item{vars}{variables}
}
\description{
Drop variables from then dataset
}
