% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{extract_more_metrics.Rautomlmetric2}
\alias{extract_more_metrics.Rautomlmetric2}
\title{Extract sprcific metric for a particular model}
\usage{
\method{extract_more_metrics}{Rautomlmetric2}(object, model_name, metric_name)
}
\arguments{
\item{object}{metic object of class Rautomlmetric2}
}
\description{
Extract sprcific metric for a particular model
}
