% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extract_more_metrics}
\alias{extract_more_metrics}
\title{Extract sprcific metric for a particular model}
\usage{
extract_more_metrics(object, model_name, metric_name)
}
\description{
Extract sprcific metric for a particular model
}
