% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utilities.R
\name{extract_recipe_text}
\alias{extract_recipe_text}
\title{Extract text from recipe oabject}
\usage{
extract_recipe_text(recipes)
}
\arguments{
\item{recipe}{recipe object}
}
\value{
a vector
}
\description{
Extract text from recipe oabject
}
