% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_shap_values.R
\name{extract_shap}
\alias{extract_shap}
\title{Extract SHAP measures from compute_shap_single}
\usage{
extract_shap(x, top_n_rank, total_n_rank)
}
\arguments{
\item{x}{compute_shap_single object}

\item{top_n_rank}{the top ranking variables, the rest are categorized as others.}
}
\description{
Extract SHAP measures from compute_shap_single
}
