% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{extract_summary.list}
\alias{extract_summary.list}
\title{Collect model summary}
\usage{
\method{extract_summary}{list}(models, summary_fun = quantile_summary)
}
\arguments{
\item{models}{Models from caretList}

\item{summary_fun}{a function to summarize metrics}
}
\description{
Collect model summary
}
\details{
This function extracts and summarizes model estimates from careList object.
}
