% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{extract_summary}
\alias{extract_summary}
\title{Collect model summary}
\usage{
extract_summary(models, summary_fun)
}
\arguments{
\item{models}{Models from caretList}

\item{summary_fun}{a function to summarize metrics}
}
\description{
Collect model summary
}
\details{
This function extracts and summarizes model estimates from careList object.
}
