% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{extract_value_labels}
\alias{extract_value_labels}
\title{Get value labels}
\usage{
extract_value_labels(df, var)
}
\arguments{
\item{df}{data.frame object.}

\item{var}{variable name.}
}
\value{
a vector.
}
\description{
Extract unique value labels
}
