% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{factor_numeric}
\alias{factor_numeric}
\title{Convert factors to numeric}
\usage{
factor_numeric(x)
}
\arguments{
\item{x}{input vector.}
}
\value{
a vector of class numeric.
}
\description{
Convert factors to numeric
}
