% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_current_data}
\alias{filter_current_data}
\title{Filter current dataset}
\usage{
filter_current_data(df, dataset_id_)
}
\description{
Filter current dataset
}
\details{
Filter a metrics df based on current data
}
