% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_data}
\alias{filter_data}
\title{Filter data}
\usage{
filter_data(df, pattern)
}
\arguments{
\item{df}{data frame.}

\item{pattern}{filter pattern.}
}
\value{
data.frame object.
}
\description{
Applies filter patterns to the data.
}
