% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_session_metric}
\alias{filter_session_metric}
\title{Filter session and metric}
\usage{
filter_session_metric(df, metric_name, session_name_, model_name)
}
\description{
Filter session and metric
}
\details{
Filter a metrics df based on session_name and metric
}
