% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{format_date_time}
\alias{format_date_time}
\title{Format data-time}
\usage{
format_date_time(x, format = "\%d-\%m-\%Y \%H:\%M:\%S")
}
\arguments{
\item{x}{date vector.}

\item{format}{date-time formate to output.}
}
\description{
Customize data-time format for file naming
}
