% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{generate_data_summary}
\alias{generate_data_summary}
\title{Generate quick data summary}
\usage{
generate_data_summary(df)
}
\arguments{
\item{df}{data.frame object.}
}
\value{
summary statistics.
}
\description{
Generate data summary statistics for all the variables.
}
