% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_data_class}
\alias{get_data_class}
\title{Get data class}
\usage{
get_data_class(path)
}
\arguments{
\item{path}{data path or url.}
}
\value{
A character string specifying the class of the data.
}
\description{
Assigns a class to the upload path or url based on the file extension.
}
