% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_endpoint_data}
\alias{get_endpoint_data}
\title{Get endpoint objects or data}
\usage{
get_endpoint_data(url, endpoint = c("ping", "metadata", "prototype"))
}
\description{
Get endpoint objects or data
}
