% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_file_ext}
\alias{get_file_ext}
\title{Get file extension}
\usage{
get_file_ext(file)
}
\arguments{
\item{file}{file name or file path with a .ext.}
}
\value{
a character specifying file extension.
}
\description{
Get file extension
}
\details{
Given a file name/path, extract the file extension.
}
