% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_file_name}
\alias{get_file_name}
\title{Get the file name without the extension}
\usage{
get_file_name(file)
}
\arguments{
\item{file}{file name or file path with a .ext.}
}
\value{
a character specifying the file name.
}
\description{
Get the file name without the extension
}
\details{
Given a file name/path, extract the file extension.
}
