% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_metrics_names.Rautomlmetric2}
\alias{get_metrics_names.Rautomlmetric2}
\title{Get trained model metrics}
\usage{
\method{get_metrics_names}{Rautomlmetric2}(object)
}
\arguments{
\item{object}{metric object of class Rautomlmetric2}
}
\description{
Get trained model metrics
}
