% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_nlevels}
\alias{get_nlevels}
\title{Get nlevels of a factor or character variable}
\usage{
get_nlevels(df = NULL, var)
}
\arguments{
\item{df}{data.frame}

\item{var}{variable}
}
\description{
Get nlevels of a factor or character variable
}
