% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{get_post_metrics}
\alias{get_post_metrics}
\title{Generate confussion matrix and variable importance}
\usage{
get_post_metrics(model, outcome, df = NULL, task = NULL)
}
\description{
Generate confussion matrix and variable importance
}
