% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_range}
\alias{get_range}
\title{Get range}
\usage{
get_range(df, var)
}
\arguments{
\item{df}{a data.frame.}

\item{var}{variable of interest}
}
\value{
vector of min and max.
}
\description{
Compute the range for numeric values.
}
