% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tables_from_schema_cohort.R
\name{get_tables_from_schema_cohort}
\alias{get_tables_from_schema_cohort}
\title{Get Tables from a Specific Schema}
\usage{
get_tables_from_schema_cohort(conn, schema_name)
}
\arguments{
\item{conn}{A database connection object created by \code{DatabaseConnector::connect()} or \code{DBI::dbConnect()}.}

\item{schema_name}{A character string specifying the schema name.}
}
\value{
A data frame of table names in the specified schema.
}
\description{
This function queries the database for all base table names in a given schema.
}
\examples{
\dontrun{
  conn <- DatabaseConnector::connect(connectionDetails)
  tables <- get_tables_from_schema_cohort(conn, schema_name)
}
}
