% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{ggunivariate}
\alias{ggunivariate}
\title{Univariate plots}
\usage{
ggunivariate(df, vartype)
}
\arguments{
\item{df}{data frame object.}

\item{var}{variable of interest.}
}
\value{
a ggplot object.
}
\description{
Plot univariate variable for a quick visualization.
}
