% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_options.R
\name{grid_options}
\alias{grid_options}
\title{grid_options}
\usage{
grid_options(algo)
}
\arguments{
\item{algo}{character string, name of machine learning algorithm}
}
\description{
auto train GBM using pre-defined strategies
}
