% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{missing_data_plot}
\alias{missing_data_plot}
\title{Visualise missing data}
\usage{
missing_data_plot(df, ...)
}
\arguments{
\item{df}{data frame}
}
\description{
Visualise missing data
}
\details{
The function provides a summary of the missing data
}
