% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{missing_prop}
\alias{missing_prop}
\title{Proportion of missing values}
\usage{
missing_prop(df)
}
\arguments{
\item{df}{input data frame}
}
\value{
data.frame
}
\description{
Proportion of missing values
}
