% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{na_codes}
\alias{na_codes}
\title{Convert levels to NA}
\usage{
na_codes(x, ...)
}
\arguments{
\item{x}{input vector.}

\item{...}{levels to be converted.}
}
\value{
a vector.
}
\description{
Convert levels to NA
}
