% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{na_if_category_single}
\alias{na_if_category_single}
\title{Convert patterns to NA for single variable}
\usage{
na_if_category_single(df, var, labels, range = FALSE)
}
\arguments{
\item{df}{input dataframe.}

\item{var}{input variable.}
}
\value{
a dataframe.
}
\description{
Create NAs based on particular patterns.
}
