% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{na_if_numeric}
\alias{na_if_numeric}
\title{Convert numeric values to NA}
\usage{
na_if_numeric(df, var, range, apply_all = FALSE)
}
\arguments{
\item{df}{input dataframe.}

\item{var}{input variable.}
}
\value{
a dataframe.
}
\description{
Create NAs bassed on particular numeric values.
}
