% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{na_if_pattern}
\alias{na_if_pattern}
\title{Convert patterns to NA}
\usage{
na_if_pattern(df, var, pattern, replacement = NA_character_, apply_all = FALSE)
}
\arguments{
\item{df}{input dataframe.}

\item{var}{input variable.}
}
\value{
a dataframe.
}
\description{
Create NAs based on particular patterns.
}
\details{
if regex not specified, every text is replaced with pattern. Same as \code{stringr::str_replace_all}.
}
