% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{na_to_values}
\alias{na_to_values}
\title{Assign missing values new values/categories}
\usage{
na_to_values(
  df,
  var,
  na_pattern,
  new_category,
  new_numeric = NULL,
  apply_all = FALSE
)
}
\arguments{
\item{df}{input dataframe.}
}
\value{
a vector.
}
\description{
Assign missing values new values/categories
}
