% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{numeric_character}
\alias{numeric_character}
\title{Convert numeric to character}
\usage{
numeric_character(x)
}
\arguments{
\item{x}{input vector.}
}
\value{
a vector of class character.
}
\description{
Convert numeric to character
}
