% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{numeric_factor}
\alias{numeric_factor}
\title{Convert numeric to factors}
\usage{
numeric_factor(x)
}
\arguments{
\item{x}{input vector.}
}
\value{
a vector of class factor.
}
\description{
Convert numeric to factors
}
