% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot.Rautomlmetric}
\alias{plot.Rautomlmetric}
\title{Plot metrics from model training}
\usage{
\method{plot}{Rautomlmetric}(metric)
}
\arguments{
\item{metric}{}
}
\description{
Plot metrics from model training
}
