% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot.Rautomlshap}
\alias{plot.Rautomlshap}
\title{Plot variable importance, most frequently identfied variables, dependency plots, and the SHAP values plots}
\usage{
\method{plot}{Rautomlshap}(x, pos = 0.5, drop_zero = TRUE, top_n = NULL, ...)
}
\arguments{
\item{x}{compute_shap object}
}
\description{
Plot variable importance, most frequently identfied variables, dependency plots, and the SHAP values plots
}
