% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot_vardep}
\alias{plot_vardep}
\title{Variable dependency plot for compute_shap object}
\usage{
plot_vardep(x)
}
\arguments{
\item{x}{compute_shap object}
}
\description{
Variable dependency plot for compute_shap object
}
