% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot_varfreq}
\alias{plot_varfreq}
\title{Most frequently selected important variables plot for compute_shap object}
\usage{
plot_varfreq(x)
}
\arguments{
\item{x}{compute_shap object}
}
\description{
Most frequently selected important variables plot for compute_shap object
}
