% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotfuns.R
\name{plot_varimp}
\alias{plot_varimp}
\title{Variable importance plot for compute_shap object}
\usage{
plot_varimp(x, ..., pos = 0.5, drop_zero = TRUE, top_n = NULL)
}
\arguments{
\item{x}{compute_shap object}
}
\description{
Variable importance plot for compute_shap object
}
