% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{post_metrics_model_filter}
\alias{post_metrics_model_filter}
\title{Filter models to show}
\usage{
post_metrics_model_filter(object, model_name)
}
\arguments{
\item{object}{from \code{get_post_metrics}}
}
\description{
Filter models to show
}
