% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{post_model_metrics.caretEnsemble}
\alias{post_model_metrics.caretEnsemble}
\title{Generate confussion matrix and variable importance for caretEnsemble}
\usage{
\method{post_model_metrics}{caretEnsemble}(models, outcome, df = NULL, task = NULL)
}
\description{
Generate confussion matrix and variable importance for caretEnsemble
}
