% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{post_model_metrics.caretList}
\alias{post_model_metrics.caretList}
\title{Generate confussion matrix and variable importance for caretlist}
\usage{
\method{post_model_metrics}{caretList}(models, outcome, df = NULL, task = NULL)
}
\description{
Generate confussion matrix and variable importance for caretlist
}
