% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{post_model_metrics}
\alias{post_model_metrics}
\title{Generate confussion matrix and variable importance for caretEnsemble}
\usage{
post_model_metrics(models, outcome, df = NULL, task = NULL)
}
\description{
Generate confussion matrix and variable importance for caretEnsemble
}
