% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{predict_endpoint}
\alias{predict_endpoint}
\title{Use endpoint to predict}
\usage{
predict_endpoint(
  url,
  new_data,
  recipes = NULL,
  model_name = model,
  update_skip = FALSE
)
}
\description{
Use endpoint to predict
}
