% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utilities.R
\name{preprocess}
\alias{preprocess}
\title{Create directory for various artifacts}
\usage{
preprocess(
  df,
  model_form,
  outcome_var,
  corr = 0,
  impute = TRUE,
  impute_methods = c("omit", "missing_mean", "missing_median", "mode_median",
    "mode_mean", "knn", "knn_linear", "bag"),
  perform_fe = TRUE,
  perform_pca = FALSE,
  up_sample = FALSE,
  up_sample_type = c("random", "smote", "bsmote", "adasyn", "rose"),
  df_test = NULL,
  exclude = NULL
)
}
\arguments{
\item{df}{a data frame.}

\item{model_form}{model formula.}

\item{corr}{a value for the threshold of absolute correlation values. The step will try to remove the minimum number of columns so that all the resulting absolute correlations are less than this value.}

\item{impute_methods}{a character specifying how to handle missing values.}
}
\description{
Create directory for various artifacts
}
\details{
Performs a number of ML preprocessing
}
