% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{preprocess_new_data}
\alias{preprocess_new_data}
\title{Preprocess new data}
\usage{
preprocess_new_data(recipes, new_data)
}
\arguments{
\item{recipes}{preprocessed recipes object}

\item{new_data}{new data}
}
\description{
Preprocess new data
}
\details{
Preprocess new (incoming) data for prediction.
}
