% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{quantile_summary}
\alias{quantile_summary}
\title{Quantile based summary function}
\usage{
quantile_summary(
  x,
  probs = c(0.025, 0.5, 0.975),
  names = c("lower", "estimate", "upper")
)
}
\arguments{
\item{x}{vector of numbers}
}
\description{
Quantile based summary function
}
