% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{recode_value_labels}
\alias{recode_value_labels}
\title{Recode value labels}
\usage{
recode_value_labels(x, old, new)
}
\arguments{
\item{x}{a vector of the input values/variable.}

\item{old}{old labels/values to be recoded.}

\item{new}{new labels.}
}
\value{
a vector.
}
\description{
Recode value labels of a character or a factor variable.
}
