% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{relabel_missing_values}
\alias{relabel_missing_values}
\title{Create new labels for missing values}
\usage{
relabel_missing_values(x, new_value)
}
\arguments{
\item{x}{vector input.}

\item{new_value}{new category or value.}
}
\value{
a vector.
}
\description{
Create new labels for missing values. For categorical values, assigns a factor or a character label, otherwise assigns a number.
}
