% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{rename_vars}
\alias{rename_vars}
\title{Rename variables}
\usage{
rename_vars(df, old, new)
}
\arguments{
\item{df}{dataframe}

\item{old}{old variable name}

\item{new}{new variable name}
}
\description{
Rename variables in a data frame
}
