% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{save_boot_estimates}
\alias{save_boot_estimates}
\title{Save boot estimates plots}
\usage{
save_boot_estimates(
  boot_list,
  dataset_id,
  session_name,
  timestamp = Sys.time(),
  output_dir = "outputs",
  sub_dir = "test_metrics"
)
}
\description{
Save boot estimates plots
}
