% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{save_model}
\alias{save_model}
\title{Save trained model}
\usage{
save_model(model, name, folder = "models", metadata = list())
}
\arguments{
\item{model}{trained model object}

\item{name}{character specifying the model}

\item{folder}{local folder to store the models}
}
\description{
Save trained model
}
\details{
Save a trained model into local storage
}
