% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{save_post_metrics_plots}
\alias{save_post_metrics_plots}
\title{Save post model metrics objects}
\usage{
save_post_metrics_plots(
  metric_list,
  dataset_id,
  session_name,
  timestamp = Sys.time(),
  output_dir = "outputs"
)
}
\description{
Save post model metrics objects
}
