% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{save_rautoml_csv}
\alias{save_rautoml_csv}
\title{Save rautoml csv}
\usage{
save_rautoml_csv(
  object,
  name,
  dataset_id,
  session_name,
  timestamp = Sys.time(),
  output_dir = "outputs"
)
}
\description{
Save rautoml csv
}
