% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{save_rautoml_plot}
\alias{save_rautoml_plot}
\title{Save any plot object}
\usage{
save_rautoml_plot(
  objects,
  name,
  dataset_id,
  session_name,
  timestamp = Sys.time(),
  output_dir = "outputs",
  metric_type = c("training_metrics", "test_metrics")
)
}
\description{
Save any plot object
}
