% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{start_model_api}
\alias{start_model_api}
\title{Start trained model API}
\usage{
start_model_api(folder, model_name, host = "127.0.0.1", port = NULL)
}
\description{
Start trained model API
}
\details{
Creates an API to access trained model
}
