% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping_rules.R
\docType{data}
\name{stopping_rules}
\alias{stopping_rules}
\title{stopping_rules}
\format{An object of class \code{list} of length 5.}
\usage{
stopping_rules
}
\description{
list, stopping_rules object
}
\keyword{datasets}
