% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{student_t_summary}
\alias{student_t_summary}
\title{Student-t based summary function}
\usage{
student_t_summary(
  x,
  conf.level = 0.95,
  names = c("lower", "estimate", "upper")
)
}
\arguments{
\item{x}{vector of numbers}
}
\description{
Student-t based summary function
}
