% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caret_models.R
\name{train_caret_models}
\alias{train_caret_models}
\title{Train multiple caret models using caret list}
\usage{
train_caret_models(df, model_form, ctrl, model_list, metric)
}
\description{
Train multiple caret models using caret list
}
