% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_utilities.R
\name{train_test_split}
\alias{train_test_split}
\title{Simple train/test split}
\usage{
train_test_split(
  data,
  type = c("single", "time", "group", "tseries"),
  prop = 3/4,
  group = NULL,
  strata = NULL,
  breaks = 4,
  pool = 0.1,
  ...,
  lag = 0
)
}
\arguments{
\item{data}{}

\item{...}{additional args}
}
\value{
list
}
\description{
Simple train/test split
}
