% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/utilities.R
\name{upload_data}
\alias{upload_data}
\alias{upload_data.csv}
\alias{upload_data.xlsx}
\alias{upload_data.xls}
\alias{upload_data.dta}
\alias{upload_data.rda}
\alias{upload_data.rds}
\alias{upload_data.spss}
\title{Utilities to upload various data formats}
\usage{
upload_data(path)

\method{upload_data}{csv}(path)

\method{upload_data}{xlsx}(path)

\method{upload_data}{xls}(path)

\method{upload_data}{dta}(path)

\method{upload_data}{rda}(path)

\method{upload_data}{rds}(path)

\method{upload_data}{spss}(path)
}
\arguments{
\item{path}{}
}
\value{
a dataframe.

a dataframe.

a dataframe.

a dataframe.

a dataframe.

a dataframe.

a dataframe.

a dataframe.
}
\description{
This function uses the file extension to determine data type and read data appropriately.

Uploads csv data

Uploads excel data

Uploads excel data

Uploads Stata data

Uploads rda data. The rda file should be a dataframe.

Uploads rds data. The rda file should be a dataframe.

Uploads .sav or .por data.
}
