% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R, R/utilities.R
\name{write_data}
\alias{write_data}
\alias{write_data.csv}
\alias{write_data.xlsx}
\alias{write_data.dta}
\alias{write_data.rda}
\alias{write_data.rds}
\alias{write_data.spss}
\title{Utilities saves data frame into various file formats}
\usage{
write_data(path, df)

\method{write_data}{csv}(path, df)

\method{write_data}{xlsx}(path, df)

\method{write_data}{dta}(path, df)

\method{write_data}{rda}(path, df)

\method{write_data}{rds}(path, df)

\method{write_data}{spss}(path, df)
}
\arguments{
\item{path}{}

\item{df}{data frame.}
}
\value{
NULL.

NULL.

NULL.

NULL.

NULL.

NULL.

NULL.
}
\description{
This function uses the file extension to determine data type and save data appropriately into the local disk.

Writes csv data

Writes xlsx data

Writes Stata data

Writes rda data. The rda file should be a dataframe.

Writes rds dataframe.

Writes .sav or .por data.
}
