% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGBM.R
\name{autoGBM}
\alias{autoGBM}
\title{autoGBM}
\usage{
autoGBM(x, y, train_hex, valid_hex, test_hex, model_path = "./")
}
\arguments{
\item{x}{independent variables}

\item{y}{dependent variable}

\item{data_hex}{H2ODataFrame}
}
\description{
auto train GBM using pre-defined strategies
}
\examples{
library(rBayesianOptimization)
library(h2o)
library(rAutoML)
library(knitr)
model_path <- c("C:/tmp")
h2o.init()
data(churn, package = "rAutoML")
churn_hex <- as.h2o(churn)
split_hex <- h2o.splitFrame(data = churn_hex, ratios = c(0.5,0.3), seed = 1234)
train_hex <- split_hex[[1]]
valid_hex <- split_hex[[2]]
test_hex  <- split_hex[[3]]
y <- "Churn."
x <- setdiff(names(churn_hex),  c(y))
autoGBM(x, y, train_hex, valid_hex, test_hex, model_path)
knitr::kable(autoGBM_eval, format = "markdown")
knitr::kable(autoGBM_eval, format = "pandoc")
models                    auc     logloss
-----------------  ----------  ----------
H2OGBM_Default      0.9136564   0.2006527
H2OGBM_StopRules    0.8995600   0.2097254
H2OGBM_CatEncode    0.9154616   0.1566602
H2OGBM_MaxDepth     0.9166058   0.1754864
H2OGBM_Random       0.9225640   0.1522216
H2OGBM_Bayesian     0.9262237   0.1591523
}
