% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoGBM.R
\name{autoGBM}
\alias{autoGBM}
\title{autoGBM}
\usage{
autoGBM(training_frame, validation_frame, x, y)
}
\arguments{
\item{x}{independent variables}

\item{y}{dependent variable}

\item{data_hex}{H2ODataFrame}
}
\description{
auto train GBM using pre-defined strategies
}
\examples{
library(rAutoML)
library(h2o)
h2o.init()
data(churn, package = "rAutoML")
data_hex <- as.h2o(churn)
split_hex <- h2o.splitFrame(data = data_hex, ratios = c(0.5,0.3), seed = 1234)
train_hex <- split_hex[[1]]
valid_hex <- split_hex[[2]]
test_hex <- split_hex[[3]]
y = "Churn."
x = colnames(data_hex)[colnames(data_hex)!=y]
autoGBM_list <- autoGBM(training_frame=train_hex, validation_frame_valid_hex, x=x, y=y)
}
