% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{boot_estimates}
\alias{boot_estimates}
\title{Bootstrap estimate for the predictive measures}
\usage{
boot_estimates(
  model,
  df,
  outcome_var,
  problem_type,
  nreps = 100,
  type = "prob",
  model_name = NULL,
  report = c("Accuracy", "AUC", "prAUC", "Kappa", "Sensitivity", "Specificity",
    "Precision", "Recall", "F1", "Pos_Pred_Value", "Neg_Pred_Value", "Detection_Rate",
    "Balanced_Accuracy", "logLoss", "RMSE", "Rsquared", "MAE"),
  summary_fun = quantile_summary,
  save_model = FALSE,
  model_id = paste0(model_name, "-", Sys.time()),
  model_folder = "models",
  recipe_folder = "recipes",
  preprocesses = NULL
)
}
\description{
Bootstrap estimate for the predictive measures
}
